compiler: Remove call status from local state during alias analysis
Information about the aliasing state of call arguments used to be
accumulated in the local, per function, alias status map and then
merged into the global #aas{} when the function was done.

This patch changes the handling of the aliasing state of call
arguments to update the `#aas{}` directly. The change requires the
conditions for determining when a single iteration has led to changes,
to also include `#aas.call_args`.

Functionally this patch leads to no changes in the output of the
compiler. It is the first step in switching the alias analysis from a
per-function unified alias status map to a representation that can
account for different execution paths within the function.