compiler: Fix bug in `nif_start` insertion
Commit 724da95cd04d2bc1d6cb89d1988b58b2096d59cd extended the
v3_core-pass to make all functions start with a `nif_start`
instruction when the `nifs` attribute isn't present in the module but
a call to `erlang:load_nif/2` is. The `nif_start` instruction is used
as a flag, allowing later compiler passes to easily identify a NIF.

The initial implementation assumed that a function body could only be
a `#c_seq{}` or a `#c_case{}` and never a `#c_letrec{}`. This patch
extends the v3_core-pass to handle function bodies consisting of a
`#c_letrec{}`. Credits for discovering a test case triggering this
omission goes to Matthew Pope.

Closes #7409