Rework the Module:handle_event/4 caching suggestion
Implement not wrapping the fun() for handle_event_function.

Fix the handle_event_fun() type.  The suggestion's type
was that the handle_event_fun() should be either
a state enter fun() or a handle event fun(),
but it should be a fun() that is both.

Revert the callback_mode/1 type checking function and introduce
params_callback_mode/2 to do the conversion into the cached value.
By that most of callback_mode_result/6 could be reverted,
only the finishing #params{} construction needs a change.