inets: Run httpd_SUITE:httpd_1_1 mostly in parallel
The esi_propagate (used the opportunity to fix the "proagate" typo) and
esi_atom_leak tests rely on shared resources (process registration and
atom counts) and are run separately.
Additionally, the cases in the `load` function could result in internal
server errors being returned on parallel tests on less than beefy
machines, and are also run separately and sequentially.