compiler: Correct errors in alias analysis pass
The alias analysis pass uses an algorithm inspired Kotzmann and
Mössenböck's 2005 algorithm for Escape Analysis. Unfortunately the
algorithm completely fails to identify some cases of aliasing due to
they way it tracks aliasing and updates annotations.

The alias analysis is performed by traversing the functions in the
module and their code. For each operation the uniqueness and alias
status are updated and annotations are updated. The unique/aliased
status is maintained in a map which maps a variable to a either a
status or another variable. Thus constructing equivalent sets in the
same way as Kotzmann and Mössenböck. At call sites aliasing
information is fed back to the callee and the alias analysis is repeated
for the callee. For loops this works correctly, but for non-loops this
can lead to bad aliased/unique verdicts. Consider the following
example:

1: tuple_element_aliasing() ->
2:    T = make_empty_binary_tuple(),
3:    X = element(1, T),
4:    Z = <<X/binary,1:8>>,
5:    {Z, T}.

6: make_empty_binary_tuple() ->
7:    {<<>>, <<>>}.

At line 4, the algorithm will mark X as unique as there is no feedback
from line 5 where a reference to the tuple from which X is extracted
is captured in the newly constructed tuple.

The obvious solution of waiting to update alias/unique annotations
until all basic blocks of a function have been processed, avoids the
error described above, but also disables many private-append
optimizations where the broken algorithm was correct. The main reason
for this, is that the algorithm does not consider flow-control
relationships between blocks in a function, to it, all blocks are
executed even if flow control guarantees that only one of two blocks
will ever be executed.

This patch greatly strengthens the analysis done by aliasing analysis
by accounting for flow-control within functions. It also changes the
data structure used to describe the alias status of variables in order
to support incorporating flow-control information in the per function
analysis.

In short, this patch does two things, neither of which can be done in
isolation without breaking the test suite:

1) The data structure used to describe the alias status of variables,
   called a Sharing State, is changed to track three aspects:

    - A variable is either unique, aliased.

    - A unique variable can be derived, that is extracted from a term
      or constructed from other variables.

    - The relationship between the source and a derived value is
      tracked so that it is possible to find the parent of a derived
      value and likewise find all values that are derived from a
      value.

2) The basic blocks of a function are now traversed reverse post
   order. When the end of a block is reached, the current sharing
   state is propagated to the block's successors by merging the
   sharing state of all the successor block's predecessors' sharing
   states.

   When all blocks have been visited and the sharing states at the end
   of each block are known, information about aliased variables are
   propagated along the edges of the execution graph during a post
   order traversal of the basic blocks.

compiler: Change data structure describing the alias status of variables

Change the data structure describing the alias status of
variables. Instead of a map, where different keys represented
different aspects of a variable, this patch maps a variable to a
single record describing all aspects of the variable. The change
greatly speeds up the alias analysis by reducing the number of visited
entries when merging two alias states while also reducing the
complexity of the merging logic.