shell: Cleanup latin1 handling in group
PR-7211 tried to solve an issue with using bytewise (latin1)
encoding when Erlang is run as an escript. The change was
done by making all chardata sent to the group be seen as
latin1 which was incorrect.

The solution was to properly set the output device to latin1
when starting Erlang and using the file API to write to the
device.

This commit basically reverts the changes of #7211 now that we
understand how it all should work.