Quiet down ct_run startup logs
The new output looks as follows:

    ============================= common test starting =============================
    cwd: /home/user/workspace/otp/_worktree/more-silent-tests/lib/inets/make_test_dir/inets_test
    make: 1 test module(s) compiled
    collected: 1 test(s), 451 case(s) in 11 suite(s)

    Testing make_test_dir.inets_test: Starting test, 451 test cases

The new module `ct_console` is introduced to serve as the main point for
new "fancy" terminal formatting.

Note that there is some duplication going on with respect to the color
formatting from `shell_docs`. We also don't really respect any "no
color" options the user might have (yet). This should later become part
of the `ct_console` module once we've figured out a clean way to do so.

The `Testing` output and below will be worked on in a following commit.