Fix SSH recursive search with CTRL+R
`group.erl` now sends `put_expand_no_trim` when entering a history search.
`ssh_cli` does not support that message and so when entering a search when
SSH is the underlying driver, the expanded result would not be displayed
even though the search functionality would still technically work.

This adds support for that `put_expand_no_trim` message when using SSH