compiler: Prune dead variables during alias analysis
Variables which die in a basic block cannot influence the alias status
of variables in successor blocks. By pruning dead variables, which are
not part of a parent-child derivation relationship of live variables,
the size of the active sharing state is reduced. The reduction in size
speeds up subsequent `aa_merge_ss/3` operations.

Combined with improved kill-set calculation
(85454714956f54679ae75ef49d10d7992e31b9e2) and an improved data
structure for describing the alias status of
variables (c389665cc991fb8a9d123fe34dc95c12bdf1263f), this patch
provides a substantial reduction of the time required for alias
analysis. For the set of modules compiled by `scripts/diffable` the
time spent in alias analysis is reduced by approximately 55%. For the
example in Issue #7432 [1], provided by José Valim, which has a large
number of variables, the reduction is even more dramatic. The time
spent in the alias analysis pass is reduced by 97%.

[1] https://github.com/erlang/otp/issues/7432

Closes: #7432