Make shutdown connections synchronous
Use monitors to wait for workers to stop.

Use `supervisor:terminate_child/2` to terminate workers
so the supervisor does not complain on them being killed.

Change restart strategy to `brutal_kill` so the workers
are terminated as they were with `exit(Worker, kill)`.

Fix a bug in `mod_esi` that it reacts to `'EXIT'` messages
not only from the spawned ESI worker process.