erts, kernel: Add native coverage support
This implementation was inspired by and partially based on
a previous implementation by Robin Morisset <rmorisset@meta.com>.

To get an idea of the exuction overhead for using various cover modes,
I compiled the entire compiler with the line_coverage option, and then
run the test case compilation_SUITE:self_compile/1 with different
configurations. I ran each test case several times on an 2017 Intel
iMac for each configuration and noted the shortest time for each:

* +JPcover false: 11.5 sec
* +JPcover line: 11.5 sec
* +JPcover line_counters: 12.4 sec

On an M1 MacBook Pro (AAarc64) and the times were the following:

* +JPcover false: 7.2 sec
* +JPcover line: 7.2 sec
* +JPcover line_counters: 8.0 sec

At least according to this benchmark, the overhead for the `line` mode
is negligible, while the overhead for `line_counters` mode is less
than 10 percent.