compiler: Use type info when adding alias annotations
Type information is already used when determining the alias status of
arguments to operations which are transformed by the destructive
update pass (as those operations have type annotations). This patch
changes the alias-status annotation update phase to make use of type
information for all operations which have type annotations as this
gives less confusing output.