Rewrite RR cache access time update to preserve table insert order
The cache ETS table is a bag table that in itself preserves insert order,
and when access times in the table are update, it is possible to
do the insertion of RRs with updated access time in the same order,
which overall preserves insert order.

To preserve the RR order; rewrite the deduplication code to not store
the deduplicated result in a map, but instead use a map for deduplication
lookup and store the result in a list.

All this to not get in the way of DNS servers that e.g. implement load balancing
by returning RRS in some specific order.