Fix incorrect range calculation for operator `rem`
acf3116e55f238 enhanced the range calculation for `rem`, but at the
same time introduced a bug. For example:

    2 rem X

If the range for `X` was unknown, the compiler would calculate the
range for the remainder as `{2,2}`, while the correct range is
`{0,2}`. This commit corrects the range calculation for the remainder
to ensure that zero is always included in the range.