ASN.1 JER: Refactor type-information handling
This commit refactors how type information for ASN.1 is handled for
the JER back-end, and is a compatible change for applications that use
the documented API. However, for a group of ASN.1 modules that depend
on each other (such a `S1AP-PDU-Descriptions`, `S1AP-Contents`,
`S1AP-IEs`, and so on), all members of the group must be recompiled
if any one of the group members is recompiled.

The internal changes are intended to help tools that need to process
type information for ASN.1 specifications, but note that the format of
the type information may be updated in a future release. Tool makers
must be prepared to update their tools if the format changes.

To explain the changes, I will use this ASN.1 specification as an
example:

    Containing DEFINITIONS AUTOMATIC TAGS ::=
    BEGIN
      Config ::= SEQUENCE {
        a INTEGER,
        b BOOLEAN
      }

      Str ::= OCTET STRING (CONTAINING Config)
    END

The first change is that instead having one `typeinfo_Type/0` function for
each ASN.1 type in a module, there is now a single `typeinfo/1` function.

For the example, the call `'Containing':typeinfo('Str')` retrieves the
type information for the `Str` type.

The second change is that for types having a `CONTAINING` constraint,
the constraint will be preserved in the type information. Thus, the
type information returned for the `Str` type will be:

    {container,octet_string,{typeinfo,{'Containing','Config'}}}.