Combine `try_end` with `deallocate`
When `try_end` with `deallocate` are combined the clearing of the
catch on the stack can be omitted. That eliminates one instruction
for x86_64 (9 bytes) and two instructions for AArch64 (8 bytes).

Also combine `try_end` followed by `move` and `deallocate` for AArch64.

These instruction combinations are used more than 800 times in the OTP
code base.