Improve reuse of registers
Improve reuse of registers by recognizing more neutral instructions
that will not force the cache to be discarded and by updating the
cache when reading from memory.

For the x86_64 JIT, with this commit 115697 loads of BEAM registers
are avoided, compared to 88786 loads before this commit.

For the AArch64 JIT, with this commit 14938 loads of BEAM registers
are avoided, compared to 14422 loads before this commit.