AArch64: Eliminate branch before a stub section
Because of the limited reach of branch instructions, stubs that
forward branches need to be emitted at regular intervals. Each stubs
section is always preceded by a `b` instruction skipping the section.

Frquently, this `b` instruction is unreachable (for example, if the
preceding instruction is a `ret` instruction).

By keeping track of whether the current position is unreachable,
unnecessary `b` instructions can be omitted. We can also emit stubs
sections somewhat earlier than strictly necessary if the current
position is unreachable. That will reduce the number of `b`
instructions that will need to be executed.