Fix link problem with Xcode 15
The dynamic libraries for the `wx` application would fail to build
with the new linker in Xcode 15 for macOS. One workaround is to force
the old linker to be used, but that will stop working when the the old
linker is eventually retired.

It turns out that `wx` uses a very old way to create dynamic
libraries. Essentially it asks the linker to accept that there are
references to undefined symbols, trusting that they will be defined
when the dynamic library is loaded. That is done using the following
options:

    -bundle -flat_namespace -undefined suppress

The option `-undefined suppress` is deprecated by the new linker and
it will consider undefined symbols to be an error.

A more modern approaph to creating dynamic libraries is to use the the
`-bundle_loader` option to point out an executable that can be used to
satisfy references to undefined symbols:

     -bundle -bundle_loader $ERL_TOP/bin/aarch64-apple-darwin22.6.0/beam.smp

Care must be taken not to include the `-bundler_loader` option in
`configure` tests that test linking, because the `beam.smp` executable
has not been built at that time and the linker in older Xcode releases
will complain if the executable does not exist. It seems that Xcode 15
will check that the executable exists only if there are references to
undefined symbols.

While at it, update the ancient build instructions for wxWidgets.