Skip beam_ssa_alias for modules with huge functions
The implementation of the `beam_ssa_alias` pass in OTP 26 can become
extremly slow for modules containing huge functions. The `master`
branch (for the OTP 27 release next year) already contains a much
better implementation that handles huge functions. That implementation
is also very different, so we don't want to back-port it to OTP 26.

To make the compiler in OTP 26 usable for huge functions, add a
heuristic that disables the alias analysis pass for functions
exceeding a certain number of blocks.

Fixes #7667