inets: change order of application start
to maintain backwards compatibility while starting ssl behind the
scenes, we execute the inets start as last step. this means that we can
detect if the inets app was already started, as someone may be testing
this as a test case. had we use `application:ensure_all_started([inets,
ssl])`, we would have changed the external behaviour.