Preprocessor: Fail compilation for unknown functions
EEP 44 suggested extensions to the preprocessor.

In OTP 19, the `-error` and `-warning` directives were implemented.

In OTP 21, the `-if` and `-elif` directives were implemented. These
directive evaluate a condition that syntactically looks like a guard
expression. Semantically, however, there were a few differences
compared to standard guard expressions.

One difference was that there were more built-in functions available,
such as `defined(Name)` to test whether `Name` is a defined macro.
Ultimately, of the suggested built-in functions, only `defined(Name)`
was actually implemented (but it was not documented).

Another difference was that calls to unknown functions would not cause
a compilation error, but would merely fail the condition and cause the
code following the `-if` directive to be skipped. The rationale was
that one could test for new guard BIFs directly by calling them in an
`-if` condition. Thus, one could test for the presence of maps like
so:

    -if(not is_map(a)).
    %% Code using maps.
    -else.
    %% Fallback code that don't use maps.
    -endif.

The idea is that in a release with maps, `not is_map(a)` would return
`true`, and in a release without maps, the call to the unknown
function `is_map/1` would fail and the code following the `-if` would
be skipped. This is, of course, a purely hypothetical example, because
maps were introduced in OTP 17 well before `-if` could handle guard
expressions.

In practice, new guard BIFs are not introduced very often. If I
haven't missed any, the only guard BIFs introduced after OTP 21 are
`min/2` and `max/2` in OTP 26. We could attempt test whether `max/2`
works in a guard like so:

    -if(max(0, 0) =:= 0).
    check(X) when max(X, 0) =:= 0 -> negative;
    check(_) -> positive.
    -else.
    check(X) ->
        case max(X, 0) of
            0 -> negative;
            _ -> positive
        end.
    -endif.

Unsurprisingly, this code compiles and works in OTP 26.

Perhaps somewhat surprisingly, this code does not compile in OTP 25:

    t.erl:5:2: badly formed 'if'
    %    5| -if(max(0, 0) =:= 0).
    %     |  ^

The reason is that calls to known non-guard BIF functions result in a
compilation error (as described in the EEP). Of course, `min/2` and
`max/2` has existed as non-guard BIFs for a long time.

Given that the current behavior has not proven itself useful, but
instead can hide typos, this commit makes calls that are not to known
guard BIFs into compilation errors.

Fixes #7706