Add (and use) proc_lib:set(and get)_label/1
Make it possible for the user to set a process label which
can be used in tools and crash reports to identity processes
but it doesn't have to be unique, as an registered name needs to be.

The process label can any term, for example {worker, 1..N},
{pool_process, 1..N} or something entirely different to identify
process that use general code.

While at it optimize fetching process info, so we don't have to
(rpc:) call the process_info(..) several times.