compiler: Rewrite swap to move
When a simple instruction such as building a tuple or calling a
non-garbage-collecting BIF is followed by a `swap` instruction,
it is sometimes possible to replace the `swap` instruction with
a `move` placed **before** the other instruction.

Here is an example of a guard BIF call:

    {bif,self,{f,0},[],{x,2}}.
    {swap,{x,1},{x,2}}.

which can be rewritten like so:

    {move,{x,1},{x,2}}.
    {bif,self,{f,0},[],{x,1}}.

The `make_fun3` instruction for creating a fun is frequently followed
by a `swap` instruction:

    {make_fun3, ...}.
    {swap,{x,0},{x,2}}.

which can be replaced with a `move` instruction:

    {move,{x,0},{x,2}}.
    {make_fun3, ...}.

This optimization is worthwhile because a `move` instruction is
generally cheaper than a `swap` instruction.