Prevent pbkdf2_hmac function from blocking VM schedulers
PBKDF2 functions are often meant to run for 100s of milliseconds at a time.
Previously, they executed on the main schedulers, blocking them, and thus, were
unsuitable for their purpose.

Update the function to use a direct call only if the iteration count is less
than or equal to 100 [1] and the keylen is less than or equal to 64 [2].

Add a few extra tests for the sha256 variant, and some test cases for various
invalid inputs.

[1] Based on `pbkdf2_hmac(sha256, <<"password">>, <<"salt">>, 100, 32)` running
in about 75-200 micriseconds on a sampling of various architectures.

[2] 64 bytes is the length of the SHA512 HMAC, the size of the longest
currently supported PBKDF2 hash function.

Fix: https://github.com/erlang/otp/issues/7769