diffable: Add dump of cumulative timing information for all modules
This patch extends the `diffable` escript to print cumulative timing
information for all modules compiled by the script when the flag
`--time` is given.

The output format is similar to the format produced by `erlc` when the
`+time` option is used, but compared to the per-module information,
the passes are listed from slowest to fastest.