[megaco|test] Tweaked mib related test cases
A test run caused the (mib-suite) traffic test case
to enter an infinit loop. A syntax error was detected
(*not* part of the test case) and this resultet
in the test code enter a infinit loop (sending
syntax errors between (one of the) mg and mgc).
Attempted to brake out of this by cancel and then
disconnect.

A following test case failure, in another test suite,
because the nodes, with apps and config, where left running
after the failure. Causing the next test case to fail.