[erts] Fix long signal delivery time for processes in dirty run queues
OTP-18841

Commit (1) changed so that signals sent to processes scheduled for dirty
execution didn't trigger rescheduling of such processes to normal schedulers
in order to handle signals, but instead rely on the dirty signal handler
processes to take care of handling the signals for such processes. This
since a steady flow of incoming signals to processes scheduled for dirty
execution caused such processes to bounce between schedulers and could
prevent them from making progress. The dirty signal handler processes did
however not take care of the signals until the receiving processes were
selected for execution. As a result of this, if a process got stuck for a
long time in a dirty run queue due to heavy usage of dirty schedulers, it
took a long time until signals were delivered to such processes. As of this
commit, dirty signal handler processes will also take care of signal handling
for processes in dirty run queues.

(1) 53b19837b188fb2e222804089313edb334e00a1e