ssl: Refactor pre TLS-1.3 main statemachine into client/server FSMs
This refactor is made to enhance eas of code understanding.
Also cleans up dead code and removes use of explicit ?FUNCTION_NAME as it
is harder to read and messes it up for language server features. Instead
we use new assertion macro to guard for typo errors when state-name is
an argument. If it is a function call undefined function check will help.