Guard `gen_*` arguments in the API
There is also a possibility to `throw/1` from `gen:where/1`, catch
that in the `gen_*` API functions, and convert to a 'badarg' there,
but in that case we also need to fix gen_fsm to not leak a `throw/1`,
so we leave it at this.  A bad name gives a 'badarg' from
`gen:where/1`, which is not to deep below the API function.