string_SUITE: Fix error output for non-string return values
Previously, if a case failed and expected a non-string, non-tuple value,
the 'io:format/2' call in the 'test_1/5' helper would error rather than
printing because of the '~ts' control sequence. For example, writing an
incorrect case in the `length` case like so will fail:

    ?TEST("abc", [], 4)

To fix this we swap the clauses so that we use '~ts' for binaries and
lists and '~w' for everything else.