Allow reserved words as record names w/out quotes
This motivation for this comes from the Nitrogen Web Frameworks's heavy
use of records, and a clash between HTML <div> element and the Erlang
div operator.

In Nitrogen, HTML elements are represented by Erlang records. For
example: the HTML `<span>` element is `#span{}` in Nitrogen.

Logically, the heavily-used HTML `<div>` element would be represented
by `#div{}` in Nitrogen, however, that specific syntax is illegal due to
`div`'s reserved word status, and must instead be represented with
`#'div'`. This syntax, however, is awkward, and has led to a workaround
that *works*, but is itself awkward (using the term `#panel{}` instead
of `#div{}` - but this in itself leads to a semantic clash, as some
frontend HTML frameworks have their own 'panel' elements that might
ideally be abstracted into a `#panel{}` element.

But, As far as I understand, there is no potential syntax clash in
allowing the syntax `#div` ito be acceptable, and have the parser
recognize that the `div` (or any reserved word) in that `context` can
only be an atom, and would never be an operator.

So this change tweaks the grammar to recognize the circumstances of:
`#reserved_word{}`.

Further, this change does not change the way the records are defined (so
the definition must still be defined with the atom properly wrapped in
quotes (e.g. `-record('div', {a,b}).`).

This PR also adds the appropriate tests in `erl_expand_records_SUITE`,
which I wasn't sure if that was appropriate place, but it seemed the
most relevant.