erts: Share external funs globally
By storing external funs in a dedicated literal area shared by
all modules and keeping a reference to them in the export entry
they refer to, they will no longer be tied to the lifetime of
the module that created them. `make_fun/3` will also use this
reference so that dynamic creation of an external fun will also
result in a literal.

Like all other literal optimizations at the moment, this will be
lost when the term is placed into an ordinary ETS table (although
compressed retains it by happenstance), but that is a general
issue for all literals which may be fixed in a later PR.