Rework error handling and presentation
The `ErrorDescriptor` terms, argument to `erl_scan:format_error/1`,
internal to `erl_scan`, but used by test cases and maybe stuff
like the language server, those terms; they have been reworked
to make more sense.

The error presentation for unterminated strings and atoms
now points at the first content character, and the content
is presented as a regular string, regardless of what kind of
string it came from (regular, triple-quoted, sigil).
Atoms are presented as quoted atoms.

For triple-quoted strings the first content character in this
case is the first character after the start delimiter, where it is
supposed to be only white space.  This, to not second guess
what the indentation is on the first line (since we can only
know what the correct indentation is when we found the last line),
to show if there is garbage there instead, to present the content
as for a regular string starting after the start quote,
and finally to not have to change this behaviour if we one day
would allow a content indicator text after the start delimiter...

Before this change the error presentation for unterminated
strings and atoms pointed at the first start quote character.