AArch64: Correct test of int-typed return value
The function `eq()` returns an `int`, but the code generated by the
JIT for AArch64 tested the entire `x0` (ARG1) register. That happened
to work because the way `eq()` is implemented, it does not return
random garbage in the upper part of `x0`. Change the test to only
check the lower part of ARG1 in case the implementation of `eq()`
changes in the future.