[erts] Fix alias signal type check macros
These macros should ignore the extra bits in the signal tag. This since
as of the changes introduced in this branch, those are not always zero
anymore (non-zero means that a trace token is attached to the signal).

This bug has never been released, so it needs no ticket of its own.