inets: Fix httpc parallel tests failing on FreeBSD
With the parallel tests recently introduced into httpc_SUITE, multiple
clients may be waiting to be accepted (`gen_tcp:accept`) by the test
server spawned in `http_test_lib` at once. On Linux, if the backlog is
full, clients wait for a while before timing out. On FreeBSD however,
the connection is instantly rejected with `ECONNRESET`, causing most
tests in the parallel groups to fail simultaneously.

Increase the backlog from the default 5 to 100 to queue up excess
requests on FreeBSD the same way as on Linux.