add xmerl_xml_indent-module
The indent module was added to provide out of the box
indented output instead of the standard xmerl_xml-module.

The code is based on the Elixir xmerl_xml_indent-package.
That package can be found on hex.pm.

The documentation was updated to mention the module and its
use using the motorcycle-example.

Test-cases have been added to the xmerl-SUITE as requested
by OTP-maintenance team.