Report arguments in the order given
If you specify positional arguments at various levels of sub-command,
e.g. globally, then after a command, etc., the usage info listed them in
the wrong order.

This changes fixes that, updates the documentation to be explicit, and
adds a related test to check that the arguments are parsed in the
expected order.

Fixes #7934