inets: Fix httpd_serve_SUITE failing on OpenBSD
The test previously tried to connect via 0.0.0.0 and ::, which may not
work depending on the network configuration. Since this to seems to be a
Linux specific feature that was relied on on the original machine used
to implement this feature, and it so far has caused problems with
multipel platforms, change the code to simply always convert into
127.0.0.1 and ::1 on systems except from Linux to prevent compatibility
issues. (Converting for any host would fail CI for Linux).