Add support for remapping the .text segment using THP on Linux
Linux allows an application to remap its .text segment at runtime
using transparent huge pages.  To do this, an application needs to
determine the start and length of the .text segment and pass this
range as an argument to the madvise(2) system call.  There are many
techniques for doing this, the approach chosen in this change is to
parse the /proc filesystem.  An alternative would be to get this
information from the ELF header.

For this to work reliably, the start address of the text segment
should be aligned to a multiple of the size of a transparent huge
page, and the length text segment should be greater than the size of a
transparent huge page and ideally a multiple of that size.

Finally, page sizes and support for multiple page sizes varies by
architecture.  This change only supports the 64-bit x86 but, in
theory, it can be generalized to other architectures that support THP.